requirements.txt
python-poppler
Pillow
google
google-auth
google-auth-oauthlib
google-cloud-vision
google-cloud-storage
